+++ backend/src/auth/dtos/user-register.dto.ts
@@ -1,6 +1,6 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { IsEmail, IsNotEmpty, IsString } from 'class-validator';
-export class CreateUserDto {
+export class UserRegisterDto {
   @IsNotEmpty({ message: '이메일은 필수값입니다.' })
   @IsEmail({ message: '이메일의 형식이 올바르지않습니다.' })
   @ApiProperty({ example: 'test@test.com', description: '이메일' })
