+++ backend/src/users/dtos/index.ts
@@ -1,0 +1,2 @@
+export * from './user.dto';
+export * from './user-update.dto';
